src/python/connectors/requirements.txt
minio
psycopg2 ## postgres
cassandra-driver 
